Move actions to own file
